// frontend/generated/updateEmail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: updateEmail
// ====================================================

export interface updateEmail {
  /**
   * Update Email
   */
  updateEmail: string;
}

export interface updateEmailVariables {
  email: string;
}